% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSeason.R
\name{GetSeason}
\alias{GetSeason}
\title{GetSeason}
\usage{
GetSeason(
  x,
  ValuesToUse = "RawValue",
  RawValues = TRUE,
  printKW = FALSE,
  do.plot = FALSE,
  UseMidObs = TRUE,
  TimeIncrMed = TRUE,
  ...
)
}
\arguments{
\item{x}{Input data frame must contain: TimeIcnr, ValuesToUse, Centype, Censored.}

\item{ValuesToUse}{Column name with data to use in the test. (Default: "RawValue")}

\item{printKW}{Logical. If TRUE the function prints a table of the Kruskal Wallis test outputs
for all tested seasons to the console.}

\item{do.plot}{Logical. Option to return a ggplot object of the boxplots for the chosen season
(Default: TRUE).}

\item{...}{Passed to the plot function.}
}
\value{
A list containing (1) A data frame with fields as follows (in addition to those columns in x), (2) the outputs from SeasonalityTest() and optionally (3) a plot object.\tabular{ll}{
\code{Season} \tab Season string (factor) \cr
\tab \cr
}
}
\description{
GetSeason evaluates an appropriate season for the analysis by performing a
Kruskal Wallis test (non-parametric ANOVA, via the SeasonalityTest()
function) on the observations using the possible time increment as the
explanatory (categorical) variable. Possible season time increments are (1)
the time increment specified by TimeIncr, and (2) any other time increment
that is a whole multiple of TimeIncr. This function also returns the
SeasonalityTest() outputs for the selected season (i.e., the user does not
need to run SeasonalityTest() separately).
}
