% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputelower.R
\name{Impute.lower}
\alias{Impute.lower}
\title{Impute.lower}
\usage{
Impute.lower(
  x,
  ValuesToUse = "RawValue",
  forwardT = "log",
  reverseT = "exp",
  do.plot = F,
  minObsforfit = 10
)
}
\arguments{
\item{x}{Input data frame must contain: ValuesToUse, myDate, CenType}

\item{forwardT}{See NADA function ros()}

\item{reverseT}{See NADA function ros()}

\item{do.plot}{Plot the ROS model}

\item{minObsforfit}{}
}
\value{
A data frame with fields as follows (in addition to those columns in x)\tabular{ll}{
\code{i1Values} \tab Replacements for ValuesToUse \cr
\tab \cr
\code{LeftImpute} \tab Description of whether the observations were imputed or not \cr
}
}
\description{
Imputes replacement values for left-censored data using the regression on order statistics (ROS)
function from the NADA package.
}
