% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectTrendData.R
\name{InspectTrendData}
\alias{InspectTrendData}
\title{InspectTrendData}
\usage{
InspectTrendData(
  x,
  Year = "Year",
  TrendPeriod = NA,
  EndYear = NA,
  propIncrTol = 0.9,
  propYearTol = 0.9,
  ReturnALLincr = FALSE,
  AnnualOK = c("MCI"),
  TimeIncrOpts = NA,
  do.plot = FALSE,
  UseMidObs = TRUE,
  FlowtoUse = NA,
  ...
)
}
\arguments{
\item{x}{Input data frame must contain: myDate, RawData, Censored}

\item{Year}{Vector of acceptable (sub-annual) time increments. Default is NA, which uses all coded time increments c("Monthly","Bi-monthly","Quarterly","Bi-annual") – but a subset of these can alternatively be specified.}

\item{TrendPeriod}{Column name for Year data to be used}

\item{EndYear}{Sets the end year of the time series. (If NA, will be selected from data)}

\item{propIncrTol}{The proportion of time increments x years that must have observations}

\item{propYearTol}{The proportion of year that must have observations}

\item{ReturnALLincr}{Logical. If true, only returns x with an additional column indicating the time increment.}

\item{AnnualOK}{A vector of variables which are allowed to have the time increment set as annual.}

\item{TimeIncrOpts}{Vector of acceptable (sub-annual) time increments. Default is NA, which uses all coded time increments c("Monthly","Bi-monthly","Quarterly","Bi-annual") – but a subset of these can alternatively be specified.}

\item{do.plot}{Produce a plot if TRUE and ReturnAllIncr is TRUE}

\item{UseMidObs}{Default is TRUE and uses value closest to the mid-point of the season. Set to FALSE, to use median of time increments.}

\item{FlowtoUse}{Name of column of flow data (only required if relevant)}

\item{...}{Passed to plot function}
}
\value{
A data frame of x with additional columns appended. Optionally a list with an extra dataframe and optionally a plot. Data frame appended fields as follows\tabular{ll}{
\code{Incr} \tab The time increment name (one row of data per time increment) \cr
\tab \cr
\code{TrendPeriodL} \tab Length of trend period \cr
\tab \cr
\code{Nobs} \tab Total number of observations \cr
\tab \cr
\code{nYear} \tab The number of years with data in the sampling period \cr
\tab \cr
\code{propYear} \tab Proportion of trend period years with observations \cr
\tab \cr
\code{nIcnrYear} \tab Number of time increment x year with observations \cr
\tab \cr
\code{propIncrYear} \tab Proportion of total time increment x year for the trend period with observations \cr
\tab \cr
\code{propCen} \tab Proportion of observations that are censored \cr
\tab \cr
\code{nCenLevelsLT} \tab Total number of unique non-detect censor levels (data with <) \cr
\tab \cr
\code{nCenLevelsT} \tab Total number of unique detection limits (data with >) \cr
\tab \cr
\code{nFlow} \tab Proportion of observations with simultaneous flow observations \cr
}
}
\description{
InspectTrendData prepares the data for subsequent trend analysis. This involves cutting down
the dataset to a specified trend period (specified by an end year and a trend period length in
years) and adds on the time increment column for the trend analysis, which is selected as the
lowest frequency increment that that meets the minimum data requirements specified (see
Snelder et al. 2022 for guidance on minimum data requirements). The function can also provide
optional summary statistics and graphs.
}
