% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveAlphaDetect.R
\name{RemoveAlphaDetect}
\alias{RemoveAlphaDetect}
\title{RemoveAlphaDetect}
\usage{
RemoveAlphaDetect(Data, ColToUse = "Value")
}
\arguments{
\item{Data}{Dataframe containing column ColToUse}

\item{ColToUse}{Name of column in dataframe with censored observations}
}
\value{
A data frame with fields as follows (in addition to those columns in x)\tabular{ll}{
\code{RawValue} \tab Numeric face value from ColToUse \cr
\tab \cr
\code{Censored} \tab Logical: whether the observation is censored or not \cr
\tab \cr
\code{CenType} \tab Factor indicating the censor type (lt: less than; gt: greater than, not: not censored) \cr
}
}
\description{
Takes a dataframe including a column observed values that includes censored
values (specified as the prefix > or <,column is type character) and returns
face values and information about the nature of the censoring.
}
