% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonalityTest.R
\name{SeasonalityTest}
\alias{SeasonalityTest}
\title{SeasonalityTest}
\usage{
SeasonalityTest(
  x,
  ValuesToUse = "RawValue",
  HiCensor = FALSE,
  main = NULL,
  Year = "Year",
  do.plot = FALSE,
  mymain = NA,
  RawValues = TRUE,
  UseMidObs = TRUE,
  TimeIncrMed = TRUE,
  ...
)
}
\arguments{
\item{x}{Input data frame must contain, Season, ValuesToUse,Centype, Censored}

\item{ValuesToUse}{Select Column with data to use in the test}

\item{do.plot}{Produces a ggplot object in a list if TRUE}

\item{...}{Passed to plot function}
}
\value{
A data frame and (optionally) a plot. If do.plot set to TRUE then a list is returned, with the first
item the data frame and the second is the boxplot. Data frame fields as follows\tabular{ll}{
\code{Observations} \tab Number of observations \cr
\tab \cr
\code{KWStat} \tab the Kruskal-Wallis rank sum statistic \cr
\tab \cr
\code{pvalue} \tab The p-value for the test (null hypothesis is that the data is NOT seasonal) \cr
\tab \cr
\code{TimeIncr} \tab The time increment of the dataset \cr
\tab \cr
\code{Season} \tab The season (time increment) used for the seasonality test \cr
}
}
\description{
The function SeasonalityTest() tests if the data are seasonal. The SeasonalityTest() function
performs a Kruskal Wallis test (non-parametric ANOVA) on the observations using season as the
explanatory (categorical) variable. The function also optionally produces a box plot of the data
grouped by season.
}
